Move committee info to the commitees view
