+++ src/main/java/ch/wisv/payments/admin/committees/CommitteeController.java
@@ -1,21 +1,28 @@
 package ch.wisv.payments.admin.committees;
 
 import ch.wisv.payments.model.Committee;
+import ch.wisv.payments.model.Order;
+import ch.wisv.payments.rest.OrderService;
+import ch.wisv.payments.rest.OrderServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
+
 @Controller
 @RequestMapping(value = "/committees")
 public class CommitteeController {
 
     private CommitteeService committeeService;
+    private OrderService orderService;
 
     @Autowired
-    public CommitteeController(CommitteeService committeeService) {
+    public CommitteeController(CommitteeService committeeService, OrderService orderService) {
         this.committeeService = committeeService;
+        this.orderService = orderService;
     }
 
     @GetMapping
@@ -31,6 +38,22 @@ public class CommitteeController {
         committeeService.addCommittee(committee);
 
         return "redirect:/committees";
+    }
+
+    @GetMapping("/{committeeId}")
+    public String committeInfoSelected(@PathVariable int committeeId, Model model) {
+
+        Committee committee = committeeService.getCommitteeById(committeeId);
+
+        List<Order> ordersByCommittee = orderService.getOrdersByCommittee(committee);
+        model.addAttribute("selected", true);
+        model.addAttribute("committees", committeeService.getAllCommittees());
+        model.addAttribute("committee", committee);
+        model.addAttribute("orders", ordersByCommittee);
+        model.addAttribute("transactionCost", orderService.getTransactionCostByCommittee(committee));
+
+        return "committeeinfo";
+
     }
 
     @ExceptionHandler(DuplicateKeyException.class)
