+++ src/main/resources/templates/committees.html
@@ -77,6 +77,7 @@
                     <th>ID</th>
                     <th>Name</th>
                     <th>Year</th>
+                    <th>Actions</th>
                 </tr>
                 </thead>
                 <tbody>
@@ -84,6 +85,13 @@
                     <td th:text="${committee.id}">Row 1 Data 1</td>
                     <td th:text="${committee.name.getName()}">Row 1 Data 1</td>
                     <td th:text="${committee.year}">Row 1 Data 2</td>
+                    <td>
+                        <div class="btn-group" role="group" aria-label="...">
+                            <a class="btn btn-info" th:href="@{/committees/} + ${committee.id}" href="#">
+                                <span class="glyphicon glyphicon-info-sign" aria-hidden="true"></span>
+                            </a>
+                        </div>
+                    </td>
                 </tr>
                 </tbody>
             </table>
